Create Abstract Class. Extend component from abstract class. Refactor render functions.
