+++ src/components/button-more.js
@@ -1,27 +1,11 @@
-import {createElement} from "../utils";
+import AbstractComponent from './abstract-component.js';
 
 const createButtonLoadMoreTemplate = () => {
   return `<button class="load-more" type="button">load more</button>`;
 };
 
-export default class LoadMoreButton {
-  constructor() {
-    this._element = null;
-  }
-
+export default class LoadMoreButton extends AbstractComponent {
   getTemplate() {
     return createButtonLoadMoreTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
