+++ src/components/filter.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import AbstractComponent from './abstract-component.js';
 
 const createFilterItemTemplate = (filter) => {
   return `<input
@@ -21,25 +21,13 @@ export const createFilterTemplate = (filterData) => {
   </section>`;
 };
 
-export default class Filter {
+export default class Filter extends AbstractComponent {
   constructor(filterData) {
+    super();
     this._filterData = filterData;
-    this._element = null;
   }
 
   getTemplate() {
     return createFilterTemplate(this._filterData);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
