+++ src/components/task-list-empty.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const createTaskListEmptyTemplate = () => {
   return (
@@ -8,24 +8,8 @@ const createTaskListEmptyTemplate = () => {
   );
 };
 
-export default class TaskListEmpty {
-  constructor() {
-    this._element = null;
-  }
-
+export default class TaskListEmpty extends AbstractComponent {
   getTemplate() {
     return createTaskListEmptyTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
